Add pseudorandom number generator to gamestate
Randomness is currently used for the following purposes:
 * Choosing where a player spawns after they die, or when the game starts.
 * Map generation.
 * The AI player.

We could eliminate randomness from the game by making game resets and player
respawns the responsibility of the caller.  But  I think making the RNG part of
the gamestate is ultimately the right move, as the game may be extended in the
future to include randomness during runtime, e.g. randomly appearing ammo.

In the future, if networking support is implemented, all clients will need to be
seeded the same so they stay in sync (have the same maps, AI decisions, etc.).